update structural code

+++ include/dataset_reader.h
@@ -21,8 +21,8 @@ class DatasetReader {
   void load_lane_observations(const std::string &dir);
   void load_pose(const std::string &dir);
   std::string config_file_;
-  std::map<int64_t, Odometry> pose_data_;
-  std::map<int64_t, FrameObservation> frame_observations_;
-  std::map<int64_t, Odometry>::iterator data_iterator_;
+  std::map<long long, Odometry> pose_data_;
+  std::map<long long, FrameObservation> frame_observations_;
+  std::map<long long, Odometry>::iterator data_iterator_;
 };
 }  // namespace mono_lane_mapping
